requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.7.0
git-filter-repo==2.47.0
greenlet==3.2.3
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
# langchain==0.3.27
# langchain-community==0.3.27
# langchain-core==0.3.72
# langchain-groq==0.3.7
# langchain-text-splitters==0.3.9
# langsmith==0.4.10
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
#openai==1.98.0
orjson==3.11.1
outcome==1.3.0.post0
packaging==25.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
PySocks==1.7.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.42
tenacity==9.1.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
zstandard==0.23.0
flask>=2.1
#python-dotenv>=0.21
pandas>=1.5
openpyxl>=3.1
requests
playwright==1.54.0


gunicorn>=20.1